FarmConnect/requirements.txt
flask
numpy
pandas
Pillow
requests
tensorflow
torch
torchvision
gunicorn
